backend/requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==3.7.1
blis==0.7.11
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.3
cryptography==45.0.6
cycler==0.12.1
cymem==2.0.11
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
fastapi==0.104.1
ffmpeg-python==0.2.0
filelock==3.18.0
fonttools==4.60.1
fsspec==2025.7.0
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.185.0
google-auth==2.23.4
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.71.0
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.34.4
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.11.1
jiwer==4.0.0
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
llvmlite==0.44.0
marisa-trie==1.2.1
MarkupSafe==3.0.2
matplotlib==3.10.7
more-itertools==10.7.0
motor==3.3.2
mpmath==1.3.0
multidict==6.7.0
murmurhash==1.0.13
networkx==3.5
numba==0.61.2
numpy==1.26.4
openai-whisper==20250625
packaging==25.0
pandas==2.3.3
pillow==12.0.0
postgrest==2.22.0
preshed==3.0.10
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.3
pydantic_core==2.41.4
pydub==0.25.1
PyJWT==2.10.1
pymongo==4.6.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.14.1
realtime==2.22.0
regex==2025.7.34
requests==2.31.0
rsa==4.9.1
safetensors==0.6.2
seaborn==0.13.2
six==1.17.0
smart-open==6.4.0
sniffio==1.3.1
soundfile==0.13.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.27.0
storage3==2.22.0
StrEnum==0.4.15
supabase==2.22.0
supabase-auth==2.22.0
supabase-functions==2.22.0
sympy==1.14.0
thinc==8.2.5
tiktoken==0.11.0
tokenizers==0.15.2
torch==2.1.0
torchaudio==2.1.0
tqdm==4.67.1
transformers==4.35.2
typer==0.9.4
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.24.0
wasabi==1.1.3
weasel==0.3.4
websockets==15.0.1
yarl==1.22.0
